+++ src/Service/Webhook/ReviewEventService.php
@@ -10,7 +10,6 @@ use DR\Review\Entity\Review\CodeReviewer;
 use DR\Review\Entity\Revision\Revision;
 use DR\Review\Message\Review\ReviewAccepted;
 use DR\Review\Message\Review\ReviewClosed;
-use DR\Review\Message\Review\ReviewCreated;
 use DR\Review\Message\Review\ReviewOpened;
 use DR\Review\Message\Review\ReviewRejected;
 use DR\Review\Message\Review\ReviewResumed;
@@ -20,16 +19,12 @@ use DR\Review\Message\Reviewer\ReviewerStateChanged;
 use DR\Review\Message\Revision\ReviewRevisionAdded;
 use DR\Review\Message\Revision\ReviewRevisionRemoved;
 use DR\Review\Service\CodeReview\CodeReviewerStateResolver;
-use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\MessageBusInterface;
-use Symfony\Component\Messenger\Stamp\DispatchAfterCurrentBusStamp;
 
 class ReviewEventService
 {
-    public function __construct(
-        private readonly CodeReviewerStateResolver $reviewerStateResolver,
-        private readonly MessageBusInterface $bus
-    ) {
+    public function __construct(private readonly CodeReviewerStateResolver $reviewerStateResolver, private readonly MessageBusInterface $bus)
+    {
     }
 
     public function reviewerAdded(CodeReview $review, CodeReviewer $reviewer, int $byUserId, bool $added): void
@@ -105,51 +100,6 @@ class ReviewEventService
     {
         foreach ($detachedRevisions as $revision) {
             $this->bus->dispatch(new ReviewRevisionRemoved((int)$review->getId(), (int)$revision->getId(), $byUserId, $revision->getTitle()));
-        }
-    }
-
-    public function revisionAddedToReview(
-        CodeReview $review,
-        Revision $revision,
-        bool $reviewCreated,
-        ?string $reviewState,
-        string $reviewersState,
-        ?int $userId = null
-    ): void {
-        $events = [];
-
-        // create events
-        if ($reviewCreated) {
-            $events[] = new ReviewCreated((int)$review->getId(), (int)$revision->getId(), $userId);
-        }
-        if ($reviewState === CodeReviewStateType::CLOSED && $review->getState() === CodeReviewStateType::OPEN) {
-            $events[] = new ReviewOpened((int)$review->getId(), $userId);
-        }
-        if ($reviewersState !== CodeReviewerStateType::OPEN
-            && $this->reviewerStateResolver->getReviewersState($review) === CodeReviewerStateType::OPEN) {
-            $events[] = new ReviewResumed((int)$review->getId(), $userId);
-        }
-        $events[] = new ReviewRevisionAdded((int)$review->getId(), (int)$revision->getId(), $userId, $revision->getTitle());
-
-        // dispatch $events
-        foreach ($events as $event) {
-            $this->bus->dispatch(new Envelope($event))->with(new DispatchAfterCurrentBusStamp());
-        }
-    }
-
-    public function revisionRemovedFromReview(CodeReview $review, Revision $revision, ?string $reviewState): void
-    {
-        $events   = [];
-        $events[] = new ReviewRevisionRemoved((int)$review->getId(), (int)$revision->getId(), null, $revision->getTitle());
-
-        // close review event
-        if ($reviewState !== $review->getState()) {
-            $events[] = new ReviewClosed((int)$review->getId(), null);
-        }
-
-        // dispatch $events
-        foreach ($events as $event) {
-            $this->bus->dispatch(new Envelope($event))->with(new DispatchAfterCurrentBusStamp());
         }
     }
 }
